chore: Move Sign up API to SignupRepository from UserRepository
