+++ domain/src/main/java/in/koreatech/koin/domain/repository/SignupRepository.kt
@@ -1,8 +1,11 @@
 package `in`.koreatech.koin.domain.repository
 
 import `in`.koreatech.koin.domain.model.term.Term
+import `in`.koreatech.koin.domain.model.user.CodeCount
+import `in`.koreatech.koin.domain.model.user.Duplicated
 import `in`.koreatech.koin.domain.model.user.Gender
 import `in`.koreatech.koin.domain.model.user.Graduated
+import `in`.koreatech.koin.domain.model.user.Verification
 
 interface SignupRepository {
     suspend fun getPrivacyTermText(): String
@@ -24,4 +27,36 @@ interface SignupRepository {
     suspend fun getPrivacyTerm(): Term
 
     suspend fun getKoinTerm(): Term
+
+    suspend fun isUsernameDuplicatedV2(nickname: String): Duplicated
+
+    suspend fun isPhoneDuplicated(phone: String): Duplicated
+
+    suspend fun postStudentRegister(
+        name: String,
+        phoneNumber: String,
+        userId: String,
+        password: String,
+        department: String,
+        studentNumber: String,
+        gender: String,
+        email: String,
+        nickname: String
+    ): Boolean
+
+    suspend fun postGeneralRegister(
+        name: String,
+        phoneNumber: String,
+        userId: String,
+        password: String,
+        gender: String,
+        email: String,
+        nickname: String
+    ): Boolean
+
+    suspend fun sendSMS(target: String): Boolean
+
+    suspend fun verifyCertificationCode(target: String, code: String): Verification
+
+    suspend fun countSMS(target: String): Result<CodeCount>
 }
